requirements.txt
asgiref==3.6.0
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.1
django-environ==0.10.0
idna==3.4
requests==2.30.0
sqlparse==0.4.4
urllib3==1.26.6
zebedee==0.0.7